backport r7120
excise the codebase from some C99 format specifiers.
(%zu -> %lu)


git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/branches/netplay@7122 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/sound/oggvorbis.c
@@ -276,7 +276,7 @@ soundDataBuffer* sound_DecodeOggVorbis(struct OggVorbisDecoderState* decoder, si
 	buffer = malloc(bufferSize + sizeof(soundDataBuffer));
 	if (buffer == NULL)
 	{
-		debug(LOG_ERROR, "couldn't allocate memory (%zu bytes requested)", bufferSize + sizeof(soundDataBuffer));
+		debug(LOG_ERROR, "couldn't allocate memory (%lu bytes requested)", (unsigned long) bufferSize + sizeof(soundDataBuffer));
 		return NULL;
 	}
